+++ core-bundle/src/Resources/contao/widgets/PageTree.php
@@ -189,7 +189,6 @@ class PageTree extends \Widget
       $("pt_' . $this->strField . '").addEvent("click", function(e) {
         e.preventDefault();
         Backend.openModalSelector({
-          "width": 768,
           "title": "' . \StringUtil::specialchars(str_replace("'", "\\'", $GLOBALS['TL_DCA'][$this->strTable]['fields'][$this->strField]['label'][0])) . '",
           "url": this.href,
           "callback": function(table, value) {